move /api/job/console-update into a internal-api service
