+++ deploy/infrabox/templates/fluentbit/configmap.yaml
@@ -15,8 +15,8 @@ data:
         Path             /var/log/containers/*{{- template "worker_namespace" . -}}*.log
         Parser           docker
         Tag              kube.*
-        Refresh_Interval 5
-        Mem_Buf_Limit    5MB
+        Refresh_Interval 2
+        Mem_Buf_Limit    50MB
         Skip_Long_Lines  On
     [FILTER]
         Name                kubernetes
@@ -27,8 +27,8 @@ data:
     [OUTPUT]
         Name  http
         Match *
-        Host {{ .Values.host }}
-        Port {{ .Values.port }}
+        Host internal-api.{{- template "system_namespace" . -}}
+        Port 8080
         URI /api/job/consoleupdate
         Format json
         tls On
