+++ deploy/infrabox/templates/internal-api/deployment.yaml
@@ -1,0 +1,28 @@
+apiVersion: extensions/v1beta1
+kind: Deployment
+metadata:
+    name: infrabox-internal-api
+    namespace: {{ template "system_namespace" . }}
+    labels:
+        app: infrabox-internal-api
+spec:
+    replicas: {{ default 2 .Values.api.replicas }}
+    template:
+        metadata:
+            labels:
+                app: infrabox-internal-api
+        spec:
+            {{ include "imagePullSecret" . | indent 8 }}
+            containers:
+            {{ include "containers_database" . | indent 12 }}
+            -
+                name: internal-api
+                image: {{ include "image_repository" . }}/internal-api:{{ include "image_tag" . }}
+                imagePullPolicy: Always
+                env:
+                {{ include "env_database" . | indent 16 }}
+                {{ include "env_general" . | indent 16 }}
+                {{ include "env_version" . | indent 16 }}
+                {{ include "env_cluster" . | indent 16 }}
+            volumes:
+                {{ include "volumes_database" . | indent 16 }}
