+++ deploy/infrabox/templates/internal-api/service.yaml
@@ -1,0 +1,16 @@
+kind: Service
+apiVersion: v1
+metadata:
+    name: infrabox-internal-api
+    namespace: {{ template "system_namespace" . }}
+    labels:
+        app: infrabox-internal-api
+spec:
+    selector:
+        app: infrabox-internal-api
+    ports:
+    -
+        name: http-web
+        protocol: TCP
+        port: 8080
+        targetPort: 8080
