+++ src/internal-api/Dockerfile
@@ -1,0 +1,10 @@
+ARG INFRABOX_BUILD_NUMBER
+FROM quay.io/infrabox/images-base:build_$INFRABOX_BUILD_NUMBER
+
+COPY src/internal-api/server.py /server.py
+COPY src/pyinfraboxutils /pyinfraboxutils
+COPY src/pyinfrabox /pyinfrabox
+
+ENV PYTHONPATH=/
+
+CMD python /server.py
